refactor: move rendering to frontends/native, normalize NES audio in Nes::get_sample()
- Move gl_backend.rs, shader_manager.rs, input.rs from platform/rendering/
  to frontends/native/ — they are SDL2/OpenGL-specific, not platform-generic
- Remove src/platform/rendering/ module entirely
- Move NES APU normalization (APU_OUTPUT_MAX / 1.177) into Nes::get_sample()
  so platform process_sample() is system-agnostic
- Update all import paths in native frontend callers
