+++ src/frontends/native/shader_manager.rs
