--- src/platform/rendering/mod.rs
@@ -1,6 +1,0 @@
-pub mod gl_backend;
-pub mod input;
-mod shader_manager;
-
-#[allow(unused_imports)]
-pub use gl_backend::{Crosshair, GlBackend, ProcAddressLoader, RenderTarget};
